conn_pool: unifying drain interfaces
