+++ include/envoy/common/BUILD
@@ -22,6 +22,10 @@ envoy_basic_cc_library(
 envoy_cc_library(
     name = "conn_pool_interface",
     hdrs = ["conn_pool.h"],
+    deps = [
+        "//include/envoy/event:deferred_deletable",
+        "//include/envoy/upstream:upstream_interface",
+    ],
 )
 
 envoy_cc_library(
