+++ include/envoy/http/conn_pool.h
@@ -48,7 +48,7 @@ public:
 /**
  * An instance of a generic connection pool.
  */
-class Instance : public Event::DeferredDeletable {
+class Instance : public Envoy::ConnectionPool::Instance {
 public:
   ~Instance() override = default;
 
@@ -56,27 +56,6 @@ public:
    * @return Http::Protocol Reports the protocol in use by this connection pool.
    */
   virtual Http::Protocol protocol() const PURE;
-
-  /**
-   * Called when a connection pool has been drained of pending requests, busy connections, and
-   * ready connections.
-   */
-  using DrainedCb = std::function<void()>;
-
-  /**
-   * Register a callback that gets called when the connection pool is fully drained. No actual
-   * draining is done. The owner of the connection pool is responsible for not creating any
-   * new streams.
-   */
-  virtual void addDrainedCallback(DrainedCb cb) PURE;
-
-  /**
-   * Actively drain all existing connection pool connections. This method can be used in cases
-   * where the connection pool is not being destroyed, but the caller wishes to make sure that
-   * all new streams take place on a new connection. For example, when a health check failure
-   * occurs.
-   */
-  virtual void drainConnections() PURE;
 
   /**
    * Determines whether the connection pool is actively processing any requests.
@@ -101,11 +80,6 @@ public:
    */
   virtual Cancellable* newStream(Http::ResponseDecoder& response_decoder,
                                  Callbacks& callbacks) PURE;
-
-  /**
-   * @return Upstream::HostDescriptionConstSharedPtr the host for which connections are pooled.
-   */
-  virtual Upstream::HostDescriptionConstSharedPtr host() const PURE;
 };
 
 using InstancePtr = std::unique_ptr<Instance>;
